libft updates

+++ 00-libft/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: mfarhan <mfarhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/30 23:49:37 by mfarhan           #+#    #+#             */
-/*   Updated: 2025/11/03 11:21:06 by mfarhan          ###   ########.fr       */
+/*   Updated: 2025/12/06 19:51:24 by mfarhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (*s++)
